14_uptask_mern-evitar que un usuario que no tiene permisos actualice o borre proyectos

+++ 14_uptask_mern/backend/src/controllers/ProjectController.ts
@@ -38,10 +38,10 @@ export class ProjectController {
         const error = new Error("Proyecto no encontrado")
         return res.status(404).json({ error: error.message })
       }
-      if(project.manager.toString() !== req.user.id.toString()) {
-        const error = new Error('Acción no válida')
-        return res.status(404).json({error: error.message})
-    }
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error("Acción no válida")
+        return res.status(404).json({ error: error.message })
+      }
       res.json(project)
     } catch (error) {
       console.log(error)
@@ -49,7 +49,17 @@ export class ProjectController {
   }
 
   static updateProject = async (req: Request, res: Response) => {
+    const { projectId } = req.params
     try {
+      const project = await Project.findById(projectId)
+      if (!project) {
+        const error = new Error("Proyecto no encontrado")
+        return res.status(404).json({ error: error.message })
+      }
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error("Solo el manager puede actualizar un proyecto")
+        return res.status(404).json({ error: error.message })
+      }
       req.project.clientName = req.body.clientName
       req.project.projectName = req.body.projectName
       req.project.description = req.body.description
@@ -61,7 +71,17 @@ export class ProjectController {
   }
 
   static deleteProject = async (req: Request, res: Response) => {
+    const { projectId } = req.params
     try {
+      const project = await Project.findById(projectId)
+      if (!project) {
+        const error = new Error("Proyecto no encontrado")
+        return res.status(404).json({ error: error.message })
+      }
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error("Solo el manager puede eliminar un proyecto")
+        return res.status(404).json({ error: error.message })
+      }
       await req.project.deleteOne()
       res.send("Proyecto Eliminado")
     } catch (error) {
